Generates CRDs from the Pydantic models
This also refactors the operator into some more focus modules, adds project
setup, versioning, and a CLI.
